[Move] AgendaStatusのテストはAgendaStatusの実装部分に
